Refactor loaddata: Replace `objects` with `objects_in_fixture` in `if` condition
Desirability: Before this change, this was the last use of `objects`, and it's
far from its previous uses, whereas `objects_in_fixture` is still in use much
later in the function. Replacing `objects` with `objects_in_fixture` will make
it easier to factor out this portion of the function into a separate method.

Correctness: `objects` is returned from `serializers.deserialize()`, which is
[documented] as returning an iterator. Not all iterators `it` necessarily have
`bool(it) is True` if and only if `it` was empty upon creation, and some may
have been full upon creation but become empty after the `for` loop.

[documented]: https://docs.djangoproject.com/en/dev/topics/serialization/#deserializing-data