Avoid using __file__ at module level in django.views.debug
The module-level constant `CURRENT_DIR` is used only in the module itself and is
undocumented, so I'm assuming it's an obscure private symbol that no one will
miss. I've replaced it with a module-level private function
`_builtin_template_path` that refactors and centralizes finding built-in
templates for the entire module.

The one tricky part is that #32105 (GH-13531) added the `html_template_path` and
`text_template_path` attributes `django.views.debug.ExceptionReporter`. I didn't
want to disturb #32105's goal of making the template paths easily override-able,
so I avoided calling `_builtin_template_path` in the class definition by making
detecting the presence of the attributes in `__init__` and setting defaults
there. Alternatives include making the attributes properties with setters or
cached properties without setters.